Support CSP for OpenAPI UI
