+++ dmr/openapi/views/json.py
@@ -2,7 +2,7 @@ from typing import ClassVar
 
 from django.http import HttpRequest, HttpResponse
 
-from dmr.internal.json import json_dump_schema
+from dmr.openapi.dump import json_dump
 from dmr.openapi.views.base import OpenAPIView
 
 
@@ -25,7 +25,7 @@ class OpenAPIJsonView(OpenAPIView):
     def get(self, request: HttpRequest) -> HttpResponse:
         """Render the OpenAPI schema as JSON response."""
         return HttpResponse(
-            content=json_dump_schema(
+            content=json_dump(
                 self.schema.convert(skip_validation=self.skip_validation),
             ),
             content_type=self.content_type,
